split interface tests
