+++ tests/Pipaslot.Mediator.Http.Tests/Serialization/ContractSerializer_InterfaceTestBase.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Xunit;
+
+namespace Pipaslot.Mediator.Http.Tests.Serialization
+{
+    public abstract class ContractSerializer_InterfaceTestBase : ContractSerializerBaseTest
+    {
+        [Fact]
+        public void Response_ResultTypeIsInterface_WillKeepTheResultType()
+        {
+            var sut = CreateSerializer();
+            var result = new Result();
+            var responseString = sut.SerializeResponse(new MediatorResponse(true, new object[] { result }));
+            var deserialized = sut.DeserializeResponse<IResult>(responseString);
+            Assert.Equal(result.GetType(), deserialized.Result.GetType());
+        }
+
+        public interface IResult { }
+        public class Result : IResult
+        {
+            public int Index { get; set; }
+        }
+    }
+}
